requirements.txt
asgiref==3.8.1
certifi==2024.12.14
coverage==7.6.10
Django==5.1.5
djangorestframework==3.15.2
djongo==1.3.7
dnspython==2.7.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pymongo==3.11.4
pytz==2024.2
PyYAML==6.0.2
setuptools==75.8.0
sqlparse==0.5.3
uritemplate==4.1.1
